rename repositories to workspace
